move model-baking library to oven for build exclusion
