+++ tools/oven/src/FBXBaker.cpp
