+++ tools/oven/src/ModelBakingLoggingCategory.cpp
