+++ tools/oven/src/ModelBakingLoggingCategory.h
