+++ tools/oven/src/TextureBaker.cpp
