+++ tools/oven/src/TextureBaker.h
